Increment and decrement testing and bug fixes.
Implemented as the mother of all tests... this wraps the existing test
suite in a loop over a range of +/- {years, months, ...}. This also
switches the tests to use the 'spec' reporter to make it easier to spot
check.

As with all tests this revealed some issues with the parsing of dates.
Specifically, some initializations woudn't zero out the hours, minutes,
seconds, and milliseconds resulting in bad results with doing date
arithmatic.